# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Maps PDF column names to Excel column names
 */
export interface ColumnMapping {
  /** Excel column name for Prospect No */
  prospectNo: string;
  /** Excel column name for CUID */
  cuid: string;
  /** Excel column name for Tare Weight */
  tareWeight: string;
  /** Excel column name for State */
  state: string;
  /** Excel column name for Branch Code */
  branchCode: string;
  /** Excel column name for Branch Name */
  branchName: string;
}

export type PdfStylePageOrientation = typeof PdfStylePageOrientation[keyof typeof PdfStylePageOrientation];


export const PdfStylePageOrientation = {
  landscape: 'landscape',
  portrait: 'portrait',
} as const;

/**
 * PDF styling configuration
 */
export interface PdfStyle {
  pageOrientation?: PdfStylePageOrientation;
  /** Hex color for first header group (e.g. #FFFF00) */
  headerColor1?: string;
  /** Hex color for second header group (e.g. #4985E8) */
  headerColor2?: string;
  fontSize?: number;
  rowHeight?: number;
}

export interface AuditTypeConfig {
  /** Audit type code (e.g. POA, TF) */
  code: string;
  /** Human-readable label (e.g. Physical Verification) */
  label: string;
}

export interface Bank {
  id: number;
  /** Full bank name */
  name: string;
  /** Short bank identifier code */
  code: string;
  /** @nullable */
  description?: string | null;
  columnMapping: ColumnMapping;
  pdfStyle?: PdfStyle;
  auditTypes: AuditTypeConfig[];
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface BankInput {
  name: string;
  code: string;
  description?: string;
  columnMapping: ColumnMapping;
  pdfStyle?: PdfStyle;
  auditTypes: AuditTypeConfig[];
  isActive?: boolean;
}

export interface BankUpdate {
  name?: string;
  code?: string;
  description?: string;
  columnMapping?: ColumnMapping;
  pdfStyle?: PdfStyle;
  auditTypes?: AuditTypeConfig[];
  isActive?: boolean;
}

export interface GeneratedFile {
  filename: string;
  branchCode: string;
  branchName: string;
  rowCount: number;
  downloadUrl: string;
  /** @nullable */
  fileSize?: number | null;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Job {
  id: number;
  bankId: number;
  bankName: string;
  auditType: string;
  status: JobStatus;
  originalFilename: string;
  fileCount: number;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type JobDetail = Job & ({
  files?: GeneratedFile[];
  /** @nullable */
  downloadAllUrl?: string | null;
});

export interface JobInput {
  bankId: number;
  auditType: string;
}

export interface Stats {
  totalJobs: number;
  completedJobs: number;
  failedJobs: number;
  totalPdfsGenerated: number;
  totalBanks: number;
  recentJobs: Job[];
}