# lib/api-zod/src/generated/types/auditTypeConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */

export interface AuditTypeConfig {
  /** Audit type code (e.g. POA, TF) */
  code: string;
  /** Human-readable label (e.g. Physical Verification) */
  label: string;
}