# lib/api-zod/src/generated/types/bank.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */
import type { AuditTypeConfig } from './auditTypeConfig';
import type { ColumnMapping } from './columnMapping';
import type { PdfStyle } from './pdfStyle';

export interface Bank {
  id: number;
  /** Full bank name */
  name: string;
  /** Short bank identifier code */
  code: string;
  /** @nullable */
  description?: string | null;
  columnMapping: ColumnMapping;
  pdfStyle?: PdfStyle;
  auditTypes: AuditTypeConfig[];
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}