# lib/api-zod/src/generated/types/columnMapping.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */

/**
 * Maps PDF column names to Excel column names
 */
export interface ColumnMapping {
  /** Excel column name for Prospect No */
  prospectNo: string;
  /** Excel column name for CUID */
  cuid: string;
  /** Excel column name for Tare Weight */
  tareWeight: string;
  /** Excel column name for State */
  state: string;
  /** Excel column name for Branch Code */
  branchCode: string;
  /** Excel column name for Branch Name */
  branchName: string;
}