# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';

export interface Job {
  id: number;
  bankId: number;
  bankName: string;
  auditType: string;
  status: JobStatus;
  originalFilename: string;
  fileCount: number;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}