# lib/api-zod/src/generated/types/jobDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedFile } from './generatedFile';
import type { Job } from './job';

export type JobDetail = Job & ({
  files?: GeneratedFile[];
  /** @nullable */
  downloadAllUrl?: string | null;
});