# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */

export interface JobInput {
  bankId: number;
  auditType: string;
}