# lib/api-zod/src/generated/types/pdfStyle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */
import type { PdfStylePageOrientation } from './pdfStylePageOrientation';

/**
 * PDF styling configuration
 */
export interface PdfStyle {
  pageOrientation?: PdfStylePageOrientation;
  /** Hex color for first header group (e.g. #FFFF00) */
  headerColor1?: string;
  /** Hex color for second header group (e.g. #4985E8) */
  headerColor2?: string;
  fontSize?: number;
  rowHeight?: number;
}