# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PDF Generator API - Upload Excel files and generate bank branch PDFs
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';

export interface Stats {
  totalJobs: number;
  completedJobs: number;
  failedJobs: number;
  totalPdfsGenerated: number;
  totalBanks: number;
  recentJobs: Job[];
}